+++ src/components/Capture/capture.js
@@ -33,9 +33,8 @@ class Capture extends Component {
 
   componentWillUnmount () {
     this.setState({uploadFallback: false})
-    clearInterval(this.webcamChecker);
-    // TODO add this callback when implementing live capture
-    // useFullScreen(false)
+    clearInterval(this.webcamChecker)
+    this.props.useFullScreen(false)
   }
 
   componentWillReceiveProps(nextProps) {
@@ -202,10 +201,9 @@ class Capture extends Component {
     this.setState({error: null})
   }
 
-  render ({useWebcam, back, i18n, termsAccepted, ...other}) {
-    const useCapture = (!this.state.uploadFallback && useWebcam && isDesktop && this.state.hasWebcam)
-    // TODO add this callback when implementing live capture
-    // useFullScreen(useCapture)
+  render ({useWebcam, back, i18n, termsAccepted, useFullScreen, ...other}) {
+    const useCapture = (!this.state.uploadFallback && useWebcam && this.state.hasWebcam)
+    useFullScreen(useCapture && this.props.method === 'face')
     return (
       process.env.PRIVACY_FEATURE_ENABLED && !termsAccepted ?
         <PrivacyStatement {...{i18n, back, acceptTerms: this.acceptTerms, ...other}}/> :